Move compatibleNativize, and and add variants for return values & events
